+++ app/src/main/java/ly/count/android/demo/Utility.java
@@ -71,7 +71,7 @@ public class Utility {
         if (amount > 0) {
             AnotherRecursiveCall(amount - 1);
         } else {
-            Countly.sharedInstance().recordHandledException(new Exception("A handled recursive exception"));
+            Countly.sharedInstance().crashes().recordHandledException(new Exception("A handled recursive exception"));
         }
     }
 }